refactor: consolidate arg parser module and extract shared multiSelect component
Merge `shared/arg-parser.ts` into `shared/args.ts` to consolidate all CLI
argument parsing into a single module. Move `parseCliArgs` and its tests,
remove the dead `parseArrayArg` function, and clean up the stale
`#cli/shared/arg-parser` import map entry from deno.json.

Extract the duplicated ~150-line terminal multiSelect UI from both
`install.ts` and `uninstall.ts` into a shared generic component at
`cli/ui/components/multi-select.ts`. The new component is parameterized
with `MultiSelectConfig` for customizable title, subtitle, hint, and
per-element styling (checkbox, label focus/blur, description), allowing
both install and uninstall to share the same implementation while
retaining their distinct visual styles. The `MultiSelectOption` type is
now generic (`<T extends string>`) and lives in the shared component
instead of install-specific types.

Remove the duplicate `parseTargetFlag` and `TargetFlagSchema` from
`uninstall.ts`, re-exporting the canonical version from `install.ts`.
