+++ cli/shared/args.test.ts
@@ -1,7 +1,14 @@
 import { assertEquals } from "#veryfront/testing/assert.ts";
 import { describe, it } from "#veryfront/testing/bdd.ts";
 import { z } from "zod";
-import { type ArgSpec, CommonArgs, createArgParser, extractArg, extractArgs } from "./args.ts";
+import {
+  type ArgSpec,
+  CommonArgs,
+  createArgParser,
+  extractArg,
+  extractArgs,
+  parseCliArgs,
+} from "./args.ts";
 import type { ParsedArgs } from "./types.ts";
 
 function makeParsedArgs(overrides: Record<string, unknown> = {}): ParsedArgs {
@@ -168,6 +175,47 @@ describe("cli/shared/args", () => {
       assertEquals(CommonArgs.projectSlug.keys.includes("project-slug"), true);
       assertEquals(CommonArgs.projectSlug.keys.includes("project"), true);
       assertEquals(CommonArgs.projectSlug.keys.includes("p"), true);
+    });
+  });
+
+  describe("parseCliArgs", () => {
+    it("should parse positional arguments", () => {
+      assertEquals(parseCliArgs(["dev"])._[0], "dev");
+    });
+
+    it("should parse long flags with values", () => {
+      assertEquals(parseCliArgs(["--port", "8080"]).port, 8080);
+    });
+
+    it("should parse long flags with equals", () => {
+      assertEquals(parseCliArgs(["--port=3000"]).port, 3000);
+    });
+
+    it("should parse boolean flags", () => {
+      assertEquals(parseCliArgs(["--help"]).help, true);
+    });
+
+    it("should resolve short aliases", () => {
+      assertEquals(parseCliArgs(["-p", "9000"]).port, 9000);
+    });
+
+    it("should resolve -h to help", () => {
+      assertEquals(parseCliArgs(["-h"]).help, true);
+    });
+
+    it("should handle --with as array flag", () => {
+      assertEquals(parseCliArgs(["--with", "react", "--with", "tailwind"]).with, [
+        "react",
+        "tailwind",
+      ]);
+    });
+
+    it("should not set default port", () => {
+      assertEquals(parseCliArgs([]).port, undefined);
+    });
+
+    it("should convert numeric string values", () => {
+      assertEquals(parseCliArgs(["--port", "4000"]).port, 4000);
     });
   });
 });
