+++ cli/ui/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./banner.ts";
+export * from "./multi-select.ts";
 export * from "./shortcuts.ts";
 export * from "./table.ts";
